Create HBaseRecoveryScript_secure.sh
Adding a script for HBase disaster recovery. This script can be run when a customer is trying to recover old data into a new cluster after the old cluster faces some catastrophic failure but the storage account is fine. This script will change the storage account of the cluster pointing it to the old cluster's container and performs the necessary service restarts and few other steps to bring the new cluster to proper working state with the flushed data from the old cluster. Some data loss is expected (Memstore data)